애니메이터 깡통추가, 근접무기 컴포넌트 깡통 추가, 애니메이션에 필요한걸 만들어봄

+++ src/ThruthGameEngine/TypeUI.h
@@ -72,7 +72,7 @@ namespace TypeUI
 	{
 		if constexpr (std::is_same_v<T, int>)
 		{
-			return ImGui::DragInt(_name, &_val, 1.0f, _min, _max);
+			return ImGui::DragInt(_name, &_val, 1.0f, (int)_min, (int)_max);
 		}
 		else if constexpr (std::is_same_v<T, bool>)
 		{
